# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EV Showroom API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  description: string;
  imageUrl?: string | null;
}

export interface Bike {
  id: number;
  name: string;
  slug: string;
  categoryId: number;
  category: Category;
  /** Price in USD */
  price: number;
  /** Range in kilometers per charge */
  rangeKm: number;
  /** Top speed in km/h */
  topSpeedKph: number;
  /** Motor power in watts */
  motorWatts: number;
  /** Battery capacity in watt-hours */
  batteryCapacityWh: number;
  /** Weight in kilograms */
  weightKg: number;
  description: string;
  shortDescription: string;
  imageUrl?: string | null;
  galleryUrls?: string[];
  features: string[];
  isFeatured: boolean;
  inStock: boolean;
  createdAt: string;
}

export type InquiryInquiryType =
  (typeof InquiryInquiryType)[keyof typeof InquiryInquiryType];

export const InquiryInquiryType = {
  test_ride: "test_ride",
  general: "general",
  pricing: "pricing",
} as const;

export interface Inquiry {
  id: number;
  name: string;
  email: string;
  phone?: string | null;
  bikeId?: number | null;
  inquiryType: InquiryInquiryType;
  message: string;
  createdAt: string;
}

export type CreateInquiryRequestInquiryType =
  (typeof CreateInquiryRequestInquiryType)[keyof typeof CreateInquiryRequestInquiryType];

export const CreateInquiryRequestInquiryType = {
  test_ride: "test_ride",
  general: "general",
  pricing: "pricing",
} as const;

export interface CreateInquiryRequest {
  /** @minLength 1 */
  name: string;
  email: string;
  phone?: string | null;
  bikeId?: number | null;
  inquiryType: CreateInquiryRequestInquiryType;
  /** @minLength 1 */
  message: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type ListBikesParams = {
  /**
   * Filter by category ID
   */
  categoryId?: number;
  /**
   * Minimum price filter
   */
  minPrice?: number;
  /**
   * Maximum price filter
   */
  maxPrice?: number;
  /**
   * Filter to featured bikes only
   */
  featured?: boolean;
};