# lib/api-zod/src/generated/types/bike.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EV Showroom API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Category } from "./category";

export interface Bike {
  id: number;
  name: string;
  slug: string;
  categoryId: number;
  category: Category;
  /** Price in USD */
  price: number;
  /** Range in kilometers per charge */
  rangeKm: number;
  /** Top speed in km/h */
  topSpeedKph: number;
  /** Motor power in watts */
  motorWatts: number;
  /** Battery capacity in watt-hours */
  batteryCapacityWh: number;
  /** Weight in kilograms */
  weightKg: number;
  description: string;
  shortDescription: string;
  imageUrl?: string | null;
  galleryUrls?: string[];
  features: string[];
  isFeatured: boolean;
  inStock: boolean;
  createdAt: Date;
}