# lib/api-zod/src/generated/types/createInquiryRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EV Showroom API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateInquiryRequestInquiryType } from "./createInquiryRequestInquiryType";

export interface CreateInquiryRequest {
  /** @minLength 1 */
  name: string;
  email: string;
  phone?: string | null;
  bikeId?: number | null;
  inquiryType: CreateInquiryRequestInquiryType;
  /** @minLength 1 */
  message: string;
}