# lib/api-zod/src/generated/types/inquiry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EV Showroom API specification
 * OpenAPI spec version: 0.1.0
 */
import type { InquiryInquiryType } from "./inquiryInquiryType";

export interface Inquiry {
  id: number;
  name: string;
  email: string;
  phone?: string | null;
  bikeId?: number | null;
  inquiryType: InquiryInquiryType;
  message: string;
  createdAt: Date;
}